add test with missing catalog name (sb)
This test changes the catalogs-.json file to make the apim/sb catalog
the default catalog. We may change this in the future but this was an
easy way to add a test to check to see if we can match a request that
is missing the catalog/env name.