Add logic to disambiguate api in plan
This code change iterates over the list of apis declared in a plan to
find the property name used to reference the plan

For example, can be
```
apis:
  myapi_1:
    name: myapi:1.0.0
```
or
```
apis:
  myapi_1:
    id: 57447febe4b03600c2a5b489
```
When a ratelimit is applied to an operation in a plan, the myapi_1 name
is used, not myapi. This discrepancy can cause issues with the code
that fills the optimizedData datastore.