Revert urlencode for intput values in a html form
Previously, I forced the urlencode for the input values in a
html form. It's incorrect, since a browser does the urlencode
when sending the form back to az-server. I rollbacked the
change that I did before and also refined the test cases
to do the urlencode to simulate the browser's behavior.