Render login form again after failed login
In order to sync up with edge gateway, az-server needs to re-send
the default/custom login form again after failed login. Also
add the feature to define the attempt limitation and set it to
unlimited for now.

Since custom consent is not supported yet. For now, add an handler
to response with 'not supported' message. Also add a test case
to verify it.