fix platform-specific compression test failure
The problem is that gzip headers include a platform identifier.
Previously, zlib was incorrectly treating macOS as just a "generic
UNIX", but now it is actually using the correct header. The updated zlib
made its way in to recent versions of node and started uncovering a new
source of platform-specific assumptions that people didn't know they
were making.

To fix this, instead of hard-coding the expected gzip compression
result, we can just generate them on the testing platform.

This is an instance of nodejs/node#12244.

Signed-off-by: Ryan Graham <r.m.graham@gmail.com>